Refactor ColumnHelper.getColumnWidth for performance
This method contained calls to relatively-expensive methods getNumMergedRegions
and getMergedRegion within a nested loop. For very large worksheets, this could
lead to severe performance problems.
Since these methods are returning static data (in the context of this method),
move the calls out of the nested loop and call them only once [per region].

Fix for issue 52834.